gimp: update to latest development version
gimp in meta-gnome is very old and still uses gtk2. This is a bad situation
for the wayland universe since its not usable there. I updated the recipe to
the latest developement version that works under wayland. On the way i have
switched the buildsystem to meson.

This is not particularly wayland specific and should ideally be moved to to
meta-oe if its better tested.